feat: add leave balance & usage calculation with holiday API support

+++ src/main/java/pharos/groupware/service/domain/leave/dto/LeaveBalanceResDto.java
@@ -2,11 +2,13 @@ package pharos.groupware.service.domain.leave.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class LeaveBalanceResDto {
     private String typeCode;
     private String typeName;
-    private int totalDays;
-    private int usedDays;
-    private int remainingDays;
+    private BigDecimal totalDays;
+    private BigDecimal usedDays;
+    private BigDecimal remainingDays;
 }
